# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Version one of the api.",
        "title": "Mbunge App API",
        "termsOfService": "http://dashboard.mbungeapp.tech/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://dashboard.mbungeapp.tech/support",
            "email": "858wpwaweru@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "paths": {
        "/api/v1/auth/sign_in": {
            "post": {
                "description": "register user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Sign-in User",
                "parameters": [
                    {
                        "description": "254727751832",
                        "name": "phone",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok\" response.LoginResponse",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "err_code：10002 参数错误； err_code：10003 校验错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "err_code：10001 登录失败",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "err_code：20001 服务错误；err_code：20002 接口错误；err_code：20003 无数据错误；err_code：20004 数据库异常；err_code：20005 缓存异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/participation/": {
            "get": {
                "description": "get all",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "All Partipations",
                "responses": {
                    "200": {
                        "description": "ok\" []db.Participation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array"
                            }
                        }
                    },
                    "400": {
                        "description": "err_code：10002 参数错误； err_code：10003 校验错误",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "err_code：10001 登录失败",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "err_code：20001 服务错误；err_code：20002 接口错误；err_code：20003 无数据错误；err_code：20004 数据库异常；err_code：20005 缓存异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}